Status quo ante
At this point, I think that -- for _simple_ calendar events -- we are back at status quo ante, but now as an administrative LTI. (Recurring events need more attention.) Log files are generated per calendar in the `APP_PATH/logs` directory, named by `pairing_hash` (a.k.a `id`). Calendar purging has also been template and added to navigation.

+++ purge.php
@@ -1,34 +1,34 @@
 <?php
 
-require_once(__DIR__ . '/.ignore.calendar-ics-authentication.inc.php');
-require_once(__DIR__ . '/config.inc.php');
+require_once('common.inc.php');
 
-require_once(APP_PATH . '/include/page-generator.inc.php');
-require_once(APP_PATH . '/include/canvas-api.inc.php');
-require_once(APP_PATH . '/include/mysql.inc.php');
+if (isset($_REQUEST['course_url'])) {
+	$eventsApi = new CanvasPest($_SESSION['apiUrl'], $_SESSION['apiToken']);
+	
+	// TODO work nicely with the cache (purge uncached events, or only cached events, etc.)
+	
+	$events = $eventsApi->get('calendar_events',
+		array(
+			'type' => 'event',
+			'all_events' => true,
+			'context_codes[]' => preg_replace('|.*/courses/(\d+)/?.*|', "course_$1", $_REQUEST['course_url'])
+		)
+	);
+	do {
+		foreach($events as $event) {
+			$api->delete("calendar_events/{$event['id']}",
+				array(
+					'cancel_reason' => $metadata['APP_NAME'] . " course_url={$_REQUEST['course_url']}"
+				)
+			);
+		}
+	} while($events = $eventsApi->nextPage());
+	
+	$smarty->assign('content', 'Calendar purged.');
+} else {
+	$smarty->assign('content', '<form action="' . $_SERVER['PHP_SELF'] . '" method="post"><label for="course_url">Course URL <input id="course_url" name="course_url" type="text" /><input type="submit" value="Purge All Calendar Events" /></form>');
+}
 
-$eventsApi = new CanvasApiProcess(CANVAS_API_URL, CANVAS_API_TOKEN);
-$api = new CanvasApiProcess(CANVAS_API_URL, CANVAS_API_TOKEN);
-
-// TODO work nicely with the cache (purge uncached events, or only cached events, etc.)
-
-$events = $eventsApi->get('calendar_events',
-	array(
-		'type' => 'event',
-		'all_events' => true,
-		'context_codes[]' => preg_replace('|.*/courses/(\d+)/?.*|', "course_$1", $_REQUEST['course_url'])
-	)
-);
-do {
-	foreach($events as $event) {
-		$api->delete("calendar_events/{$event['id']}",
-			array(
-				'cancel_reason' => TOOL_NAME . " course_url={$_REQUEST['course_url']}"
-			)
-		);
-	}
-} while($events = $eventsApi->nextPage());
-
-echo('Calendar purged.');
+$smarty->display();
 
 ?>
